requirements.txt
pytest == 8.4.1
selene == 2.0.0rc9
requests == 2.32.4
allure-pytest